channel to event tree conversion code
Converts data stored in Nabin's root tree (one file per channel) to my root tree format as events (a one file per board type). The number of entries in the tree as obtained by tree.GetEntries() is the number of events. No external user input should be necessary, even for the part of the code where we read Nabin's tree.